move some of the buttons on profile page to v2
